Moved WrapLayout to its own public class
